chore: add coverage to transaction.go
