+++ receiver/prometheusreceiver/internal/metadata_test.go
@@ -6,34 +6,10 @@ package internal
 import (
 	"testing"
 
+	fakemetadata "github.com/open-telemetry/opentelemetry-collector-contrib/receiver/prometheusreceiver/internal/metadata"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/scrape"
 )
-
-type fakeMetadataStore struct {
-	data map[string]scrape.MetricMetadata
-}
-
-func (f fakeMetadataStore) ListMetadata() []scrape.MetricMetadata {
-	out := make([]scrape.MetricMetadata, 0, len(f.data))
-	for _, m := range f.data {
-		out = append(out, m)
-	}
-	return out
-}
-
-func (f fakeMetadataStore) GetMetadata(name string) (scrape.MetricMetadata, bool) {
-	m, ok := f.data[name]
-	return m, ok
-}
-
-func (f fakeMetadataStore) SizeMetadata() int {
-	return len(f.data)
-}
-
-func (f fakeMetadataStore) LengthMetadata() int {
-	return len(f.data)
-}
 
 func TestMetadataForMetric_Internal(t *testing.T) {
 	// Internal metric should return from internalMetricMetadata, ignoring external store.
@@ -50,15 +26,14 @@ func TestMetadataForMetric_Internal(t *testing.T) {
 }
 
 func TestMetadataForMetric_ExternalExactHit(t *testing.T) {
-	store := fakeMetadataStore{
-		data: map[string]scrape.MetricMetadata{
-			"http_requests_total": {
-				MetricFamily: "http_requests_total",
-				Type:         model.MetricTypeCounter,
-				Help:         "Total HTTP requests",
-			},
+	store := fakemetadata.NewFakeMetadataStore(map[string]scrape.MetricMetadata{
+		"http_requests_total": {
+			MetricFamily: "http_requests_total",
+			Type:         model.MetricTypeCounter,
+			Help:         "Total HTTP requests",
 		},
-	}
+	},
+	)
 	metadata, resolved := metadataForMetric("http_requests_total", store)
 	if resolved != "http_requests_total" {
 		t.Fatalf("expected resolved name http_requests_total, got %q", resolved)
@@ -70,15 +45,13 @@ func TestMetadataForMetric_ExternalExactHit(t *testing.T) {
 
 func TestMetadataForMetric_NormalizedFallback_Gauge(t *testing.T) {
 	// Simulate a merged metric like "histogram_count" where store only knows the base name.
-	store := fakeMetadataStore{
-		data: map[string]scrape.MetricMetadata{
-			"histogram": {
-				MetricFamily: "histogram",
-				Type:         model.MetricTypeGauge,
-				Help:         "Histogram base metric",
-			},
+	store := fakemetadata.NewFakeMetadataStore(map[string]scrape.MetricMetadata{
+		"histogram": {
+			MetricFamily: "histogram",
+			Type:         model.MetricTypeGauge,
+			Help:         "Histogram base metric",
 		},
-	}
+	})
 	metadata, resolved := metadataForMetric("histogram_count", store)
 	// For non-counter types, resolved should be the normalized base name.
 	if resolved != "histogram" {
@@ -94,15 +67,14 @@ func TestMetadataForMetric_NormalizedFallback_Gauge(t *testing.T) {
 
 func TestMetadataForMetric_NormalizedFallback_CounterKeepsOriginal(t *testing.T) {
 	// If normalized metadata type is Counter, resolved should stay the original name.
-	store := fakeMetadataStore{
-		data: map[string]scrape.MetricMetadata{
-			"requests": {
-				MetricFamily: "requests",
-				Type:         model.MetricTypeCounter,
-				Help:         "Requests counter",
-			},
+	store := fakemetadata.NewFakeMetadataStore(map[string]scrape.MetricMetadata{
+		"requests": {
+			MetricFamily: "requests",
+			Type:         model.MetricTypeCounter,
+			Help:         "Requests counter",
 		},
-	}
+	},
+	)
 	metadata, resolved := metadataForMetric("requests_total", store)
 	if resolved != "requests_total" {
 		t.Fatalf("expected resolved name requests_total, got %q", resolved)
@@ -133,15 +105,14 @@ func TestMetadataForMetric_Unknown(t *testing.T) {
 
 func TestMetadataForMetric_PrefersInternalOverExternal(t *testing.T) {
 	// Ensure internal metric metadata is used even if external store provides a conflicting entry.
-	store := fakeMetadataStore{
-		data: map[string]scrape.MetricMetadata{
-			scrapeUpMetricName: {
-				MetricFamily: "up_external",
-				Type:         model.MetricTypeCounter,
-				Help:         "External override attempt",
-			},
+	store := fakemetadata.NewFakeMetadataStore(map[string]scrape.MetricMetadata{
+		scrapeUpMetricName: {
+			MetricFamily: "up_external",
+			Type:         model.MetricTypeCounter,
+			Help:         "External override attempt",
 		},
-	}
+	},
+	)
 	m, resolved := metadataForMetric(scrapeUpMetricName, store)
 	if resolved != scrapeUpMetricName {
 		t.Fatalf("expected resolved name %q, got %q", scrapeUpMetricName, resolved)
